feat(apply): share one timeline renderer between both surfaces
admin/apply_live.ex had its own phase_label/1, duplicating what
region_tab.ex used to have as phase_text/1. Extract the shared markup
into AtlasWeb.Components.ApplyTimelineComponent.timeline/1 and have both
region_tab.ex and apply_live.ex render it from the same
Atlas.Control.ApplyTimeline broadcast, removing the duplication.
